implementing createPgDbApi
